renamed material layout files
