+++ packages/material/src/layouts/index.ts
@@ -1,3 +1,3 @@
 export * from './group.layout';
-export * from './material-horizontal.layout';
-export * from './vertical.layout';
+export * from './MaterialHorizontalLayout';
+export * from './MaterialVerticalLayout';
